requirements.txt
fastapi
uvicorn
httpie
requests
httpx
jinja2
python-multipart
pillow
bcrypt
python-dotenv
mercadopago
pyJWT
python-dotenv